ocr-service/pyproject.toml
[build-system]
requires = ["setuptools>=68.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ocr-service"
version = "0.1.0"
description = "Modular OCR service with pluggable engine support for LibreChat"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.109.0",
    "uvicorn[standard]==0.27.0",
    "python-multipart==0.0.6",
    "pydantic==2.6.0",
    "pydantic-settings==2.1.0",
    "pytesseract==0.3.10",
    "Pillow==10.2.0",
    "pdf2image==1.17.0",
    "python-magic==0.4.27",
]

[project.optional-dependencies]
dev = [
    "pytest==8.0.0",
    "pytest-asyncio==0.23.4",
    "httpx==0.26.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

